refactor(daemon): break up irrlichd main() into named startup phases
CodeScene flagged core/cmd/irrlichd/main.go as the repo's top refactoring
target (Code Health 5.16/10, 106 revisions, 844 LOC). The root cause: a
~740-line main() doing CLI dispatch, config, and the full daemon wiring
(relay, HTTP routes, process watcher, detector, permissions, backchannel,
sockets, mDNS, shutdown) inline in one function with deep nesting and many
forward-referenced closures — exactly the kind of complexity that let #795's
route-registration race hide in plain sight.

Extracts each cohesive phase into a named, independently readable function
(setupRelay, setupProcessWatcher, registerCoreRoutes, buildDetector,
setupPermissionService, setupBackchannel, startBackgroundLoops, shutdown,
etc.) in a new startup.go, moves the path/diagnostics helpers into paths.go,
and relocates two handler-only constants into handlers.go. main() now only
sequences these calls; runDaemon() replaces the giant inline body. No
behavior change — same construction order, same defer/shutdown sequence,
same env vars and routes.
